# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "My API",
        "description": "REST API for the sonotrak application."
    },
...
    "host": "localhost:8080",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "definitions": {
...
        "route": {
            "type": "object",
            "properties": {
                "agency_id": {
                    "type": "string"
                },
                "route_short_name": {
                    "type": "string"
                },
                "route_long_name": {
                    "type": "string"
                },
                "route_desc": {
                    "type": "string"
                },
                "route_type": {
                    "type": "integer"
                },
                "route_url": {
                    "type": "string"
                },
                "route_color": {
                    "type": "string"
                },
                "route_text_color": {
                    "type": "string"
                }
            }
        },
        "trip": {
            "type": "object",
            "properties": {
                "route_id": {
                    "type": "string"
                },
                "service_id": {
                    "type": "string"
                },
                "trip_headsign": {
                    "type": "string"
                },
                "trip_short_name": {
                    "type": "string"
                },
                "direction_id": {
                    "type": "integer"
                },
                "block_id": {
                    "type": "string"
                },
                "vehicle_id": {
                    "type": "string"
                },
                "schedule_relationship": {
                    "type": "string",
                    "enum": [
                        "SCHEDULED",
                        "CANCELED"
                    ]
                }
            }
        },
...
    },
    "paths": {
...
        "/api/route/{id}": {
            "get": {
                "summary": "get route by id ",
                "description": "Updates an existing route with new information",
                "tags": [
                    "Route"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/route"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            },
            "put": {
                "summary": "Update an existing route",
                "description": "Updates an existing route with new information",
                "tags": [
                    "Route"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/route"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/route"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            },
            "delete": {
                "summary": "delete route by id ",
                "description": "delete an existing route",
                "tags": [
                    "Route"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/route"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            }
        },
...
        "/api/trip/{id}": {
            "get": {
                "summary": "get trip by id ",
                "description": "Updates an existing trip with new information",
                "tags": [
                    "Trip"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/trip"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            },
            "put": {
                "summary": "Update an existing trip",
                "description": "Updates an existing trip with new information",
                "tags": [
                    "Trip"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/trip"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/trip"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            },
            "delete": {
                "summary": "delete trip by id ",
                "description": "delete an existing trip",
                "tags": [
                    "Trip"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/trip"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            }
        },
...
        "/api/timetable/{agency}": {
            "get": {
                "summary": "get timetable by agency ",
                "description": "get timetable for specific agency .",
                "tags": [
                    "Timetable"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "agency",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/timetable/{agency}/{stop}": {
            "get": {
                "summary": "get timetable by agency for specific stop ",
                "description": "get timetable for specific agency  for specifc stop.",
                "tags": [
                    "Timetable"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "agency",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "stop",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    }
}